Relocate stocktake tests to graphql crate
