+++ graphql/stocktake/Cargo.toml
@@ -25,6 +25,11 @@ serde = "1.0.126"
 serde_json = "1.0.66"
 thiserror = "1.0.30"
 
+[dev-dependencies]
+actix-rt = "1.1.1" # for Tokio 0.2
+assert-json-diff = "2.0.1"
+
+
 [features]
 default = ["repository/sqlite"]
 postgres = ["repository/postgres"]
